[IMPROVEMENT](#794)Refactoring beautify related code for code-editor [DESC.] * Refactoring beautify related code for code-editor into a new beautify content assist plugin
