+++ apps/ide/src/plugins/webida.editor.code-editor.content-assist.beautify/plugin.js
@@ -1,0 +1,27 @@
+/*
+ * Copyright (c) 2012-2015 S-Core Co., Ltd.
+ * 
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/**
+ * This plug-in manages life-cycle of the beautify content assist plugin
+ * 
+ * @since: 2015.12.02
+ * @author: h.m.kwon
+ */
+define([], function () {
+    'use strict';
+
+    //TODO : Plugin should be used as a LifeCycle Management
+});
